move GatewayIntents to model
