+++ src/client/bridge/gateway/shard_manager.rs
@@ -11,7 +11,6 @@ use tracing::{info, instrument, warn};
 use typemap_rev::TypeMap;
 
 use super::{
-    GatewayIntents,
     ShardId,
     ShardManagerMessage,
     ShardManagerMonitor,
@@ -27,6 +26,7 @@ use crate::framework::Framework;
 use crate::internal::prelude::*;
 use crate::internal::tokio::spawn_named;
 use crate::CacheAndHttp;
+use crate::model::gateway::GatewayIntents;
 
 /// A manager for handling the status of shards by starting them, restarting
 /// them, and stopping them when required.
@@ -53,7 +53,8 @@ use crate::CacheAndHttp;
 /// use std::env;
 /// use std::sync::Arc;
 ///
-/// use serenity::client::bridge::gateway::{GatewayIntents, ShardManager, ShardManagerOptions};
+/// use serenity::model::gateway::GatewayIntents;
+/// use serenity::client::bridge::gateway::{ShardManager, ShardManagerOptions};
 /// use serenity::client::{EventHandler, RawEventHandler};
 /// use serenity::framework::{Framework, StandardFramework};
 /// use serenity::http::Http;
